Change code_dir and live_dir locations to work with dev container
Currently  the code_dir is set to /opt/code and the live_dir is set to /code.
Problems arise when virtual environments like python's pip or bin stubs
are needed to be created within the code_dir during the build process.
Scripts that are generated will be shebanged to /opt/code instead of code,
which when placed in the dev container, will actually exist at /code and
will thus break.

All of the hooks already reference the constants, so this change should
be trivial and without consequence.